feat(purchase order): create purchase order when manager confirm purchase request

+++ src/main/java/fpt/edu/vn/gms/repository/PurchaseOrderRepository.java
@@ -3,5 +3,5 @@ package fpt.edu.vn.gms.repository;
 import fpt.edu.vn.gms.entity.PurchaseOrder;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PurchaseOrderRepository extends JpaRepository<PurchaseOrder, Integer> {
+public interface PurchaseOrderRepository extends JpaRepository<PurchaseOrder, Long> {
 }
